done with comments routes
