+++ backend/app/api/comment_routes.py
@@ -1,44 +1,12 @@
-from flask import Blueprint, request, redirect, url_for
+from flask import Blueprint, request, redirect
 from flask_login import login_required, current_user
-from app.models import Comment, db, Post, User, Follow, Reply
+from app.models import Comment, Post, Reply, db
 from app.forms import CommentForm, ReplyForm
 from .post_routes import authorized_follower
 from .auth_routes import validation_errors_to_error_messages
 
 
 comment_routes = Blueprint("comments", __name__)
-
-# Modify it in post routes
-def authorized_follower2(cb):
-    """
-    Check's if:
-        - Post belongs to a Public User or
-        - Current user is a follower or
-        - Post belongs to current user
-    """
-    def wrapper(*args, **kwargs):
-
-        post_id = kwargs.get("post_id")
-        if "comment_id" in kwargs:
-            comment = Comment.query.get_or_404(kwargs["comment_id"])
-            print(comment.to_dict())
-            post_id = comment.post_id
-            print(post_id)
-
-        post = Post.query.get_or_404(post_id)
-        print(post.to_dict())
-        owner = User.query.get_or_404(post.user_id)
-        is_owner = owner.id == current_user.id
-
-        if not is_owner and owner.is_private and not post.is_story:
-            follow = Follow.query.filter_by(
-                follower_id=current_user.id, following_id=post.user_id, is_pending=False).first()
-            if not follow:
-                return redirect(url_for("auth.unauthorized"))
-
-        return cb(kwargs.get("post_id") or kwargs.get("comment_id"))
-    wrapper.__name__ = cb.__name__
-    return wrapper
 
 
 @comment_routes.route("/<int:comment_id>", methods=["PUT"])
@@ -79,49 +47,9 @@ def delete_comment(comment_id):
     return redirect("../auth/unauthorized")
 
 
-# Move to post routes
-@comment_routes.route("/<int:post_id>/comments", methods=["GET"])
-@login_required
-@authorized_follower2
-def get_comments(post_id):
-    """
-    Query for all comments of a post
-    """
-    post = Post.query.get_or_404(post_id)
-    comments = post.comments
-    return {"Comments": [comment.to_dict() for comment in comments]}
-
-
-# Move to post routes
-@comment_routes.route("/<int:post_id>/comments", methods=["POST"])
-@login_required
-@authorized_follower2
-def add_comment(post_id):
-    """
-    Query for a post by id
-
-    Creates a comment on the post
-    """
-    form = CommentForm()
-    post = Post.query.get(post_id)
-
-    form["csrf_token"].data = request.cookies["csrf_token"]
-    if form.validate_on_submit():
-        comment = Comment(
-            user_id = current_user.id,
-            post_id = post.id,
-            comment = form.data["comment"]
-        )
-        db.session.add(comment)
-        db.session.commit()
-        return comment.to_dict()
-    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
-
-
-# Go back to authorized_follower func in post routes
 @comment_routes.route("/<int:comment_id>/replies", methods=["POST"])
 @login_required
-@authorized_follower2
+@authorized_follower
 def add_reply(comment_id):
     """
     Query for a comment by id
@@ -146,7 +74,7 @@ def add_reply(comment_id):
 
 @comment_routes.route("/<int:comment_id>/replies", methods=["GET"])
 @login_required
-@authorized_follower2
+@authorized_follower
 def get_replies(comment_id):
     """
     Query for all replies of a comment
